Split large components into focused, single-responsibility files
AdminEventsManager (730 lines) split into three:
- EditEventForm: owns edit state, image upload, and save logic
- EventCard: owns the read view and delete confirmation UI
- AdminEventsManager: lean list manager (~115 lines)

Menu: converted renderFilterSection helper function into a proper
FilterSection component, making the structure explicit and testable.
